chore: cleanup tests and database management

+++ api/app/database/__init__.py
@@ -21,14 +21,12 @@ def init_db(app: Flask) -> None:
     -------
     None
     """
-    db_path = app.config["DATABASE_PATH"]
-    if os.path.isfile(db_path):
-        os.remove(db_path)
-    with open(db_path, "w+") as f:
-        f.write("")
+    if app.config["APP_ENV"] != "testing":
+        db_path = app.config["DATABASE_PATH"]
+        if os.path.isfile(db_path):
+            os.remove(db_path)
+        with open(db_path, "w+") as f:
+            f.write("")
     from app.models import Base
-    Database.init(
-        db_uri=app.config["DATABASE_URI"],
-        db_path=db_path
-    )
+    Database.init(db_uri=app.config["DATABASE_URI"])
     Base.metadata.create_all(Database.get_engine())
